+++ arduino/scout/NewTone.h
@@ -11,7 +11,7 @@ const uint8_t SPEAKER_B_MASK = 0x04;
 const uint8_t SPEAKER_C_MASK = 0x08;
 const uint8_t SPEAKER_D_MASK = 0x10;
 
-extern uint8_t speaker_preload[BUFFER_MAX];
+extern uint16_t speaker_preload[BUFFER_MAX];
 
 void newToneSetup();
